requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
deepseek-ai==0.0.2
discord-webhook==1.3.0 